refactor: update block catalog tests to check for margin styles and add new storage tests
